% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBS.R
\name{convert_BS_timeseries}
\alias{convert_BS_timeseries}
\title{convert_BS_timeseries}
\usage{
convert_BS_timeseries(in_files, in_dates, out_rast = out_rast, shp_in)
}
\arguments{
\item{shp_in}{}
}
\description{
convert_BS_timeseries
}

