% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fishnet.R
\name{create_fishnet}
\alias{create_fishnet}
\title{create_fishnet}
\usage{
create_fishnet(in_ext = NULL, in_rast = NULL, out_name_shp = NULL,
  cellsize = NULL, in_proj)
}
\arguments{
\item{in_rast}{}

\item{out_name_shp}{string(optional)}
}
\description{
funtion to create a polygon fishnet of specified resoultion over a spatial object
 Object can be a raster file, an extent, or a generic *SP object. If a raster is selected, the
 grid is made so that upper left corner coincides with upper left corner of the raster.
If both in_ext and in_rast are declared, the fishnet is cropped on raster extent.
}

