% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datesconversion.R
\name{datetodoy}
\alias{datetodoy}
\title{Title datetodoy}
\usage{
datetodoy(dates = dates)
}
\arguments{
\item{dates}{dates to be converted. class = date, posIx, etc....}
}
\value{
doy / array of doys
}
\description{
Converts a date (or array of dates) to doy
}

