% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datesconversion.R
\name{doytodate}
\alias{doytodate}
\title{Title doytodate}
\usage{
doytodate(doys = doys, year = NULL)
}
\arguments{
\item{doys}{to be converted}

\item{year}{year to be used as basis (careful with LEAP years !!!)}
}
\value{
date / array of dates
}
\description{
Converts a doy (or array of doys) to date format. Year has to be specified
}

