% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent_reproj.R
\name{extent_reproj}
\alias{extent_reproj}
\title{extent_reproj}
\usage{
extent_reproj(ext_in, in_proj, out_proj)
}
\arguments{
\item{ext_in}{"Extent" input extent}

\item{in_proj}{"proj4string" projection of the input extent}

\item{out_proj}{"proj4string" desired projection for the output extent}
}
\value{
reprojected extent (in out_proj" projection)
}
\description{
function to convert the extent of a spatial object to a different projection
}
\examples{

}

