% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_batch_clip.R
\name{lb_batch_clip}
\alias{lb_batch_clip}
\title{Title}
\usage{
lb_batch_clip(in_folder, out_folder, in_clip_shape, in_pattern = "*.tif$",
  in_nodata = NULL, out_nodata = NULL, out_format = "GTiff",
  out_proj = NULL, recursive = FALSE)
}
\arguments{
\item{recursive}{}
}
\description{
Title
}

