% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastzonal.R
\name{lb_fastzonal}
\alias{lb_fastzonal}
\title{Title}
\usage{
lb_fastzonal(in_rts, sp_object, start_date = NULL, end_date = NULL,
  id_field = NULL, FUN = "mean", out_format = "xts", small = TRUE,
  small_method = "centroids", na.rm = TRUE, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
\examples{

}

