% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_rastervarious.R
\name{lb_fuzzyconversion}
\alias{lb_fuzzyconversion}
\title{Title}
\usage{
lb_fuzzyconversion(in_raster = in_raster, thresholds = thresholds, ...)
}
\arguments{
\item{...}{any other parameter passable to "calibrate"}
}
\value{
raster object obtained applying a fuzzy "calibration" on the input according to specified thresholds.
See help for "calibrate" function in QCA library for details
}
\description{
Title
}

