% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PGIS.R
\name{lb_get_PGIS}
\alias{lb_get_PGIS}
\title{lb_get_PGIS}
\usage{
lb_get_PGIS(schema_name, table_name, con, geom_field = NULL, verbose = F)
}
\arguments{
\item{verbose}{}
}
\description{
Derived from: http://www.r-bloggers.com/load-postgis-geometries-in-r-without-rgdal
}

