% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_gettsdata.R
\name{lb_gettsdata}
\alias{lb_gettsdata}
\title{Title}
\usage{
lb_gettsdata(in_files, in_dates, shp, id_field, buffer = NULL, BS = F,
  format = "ENVI", avg = T, sd = F, minmax = F, cvar = F,
  out_rast = NULL, start_date = NULL, end_date = NULL, verbose = T,
  na.rm = T)
}
\arguments{
\item{na.rm}{}
}
\description{
Title
}

