% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_openshape.R
\name{lb_openshape}
\alias{lb_openshape}
\title{lb_openshape}
\usage{
lb_openshape(shp_file)
}
\arguments{
\item{shp_file}{filename of ESRI shapefile}
}
\value{
Spatial* object (SpatialPolygons, SpatialPoints, eccetera)
}
\description{
lb_openshape
}
\details{
function for easily opening a ESRI shapefile by simply specifying its filename
}
\examples{
\dontrun{
 shp_file = 'D:/temp/polygons.shp'
 shp = lb_openshape(shp_file)
}
}

