% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_write_envits.R
\name{lb_write_envits}
\alias{lb_write_envits}
\title{Title}
\usage{
lb_write_envits(in_raster = in_raster, in_dates = in_dates,
  out_file = out_file, date_type = "date", multiband = F)
}
\arguments{
\item{in_raster}{a}

\item{in_dates}{b}

\item{out_file}{b}

\item{date_type}{n}

\item{multiband}{n}
}
\description{
Title
}

