% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openshape.R
\name{openshape}
\alias{openshape}
\title{openshape}
\usage{
openshape(shp_file)
}
\arguments{
\item{shp_file}{filename of ESRI shapefile}
}
\value{
Spatial* object (SpatialPolygons, SpatialPoints, eccetera)
}
\description{
function for easily opening a ESRI shapefile by simply specifying its filename
}
\examples{
\dontrun{
 shp_file = 'D:/temp/polygons.shp'
 shp = lb_openshape(shp_file)
}
}

