% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeshape.R
\name{writeshape}
\alias{writeshape}
\title{writeshape}
\usage{
writeshape(out_obj, out_file, overw = FALSE)
}
\arguments{
\item{out_obj}{spatial object to save as ESRI shapefile}

\item{out_file}{filename of output ESRI shapefile}

\item{overw}{logical. if TRUE, existing files are overwritten (defaults to FALSE)}
}
\value{
NULL - out_obj is ssaved in shp_file
}
\description{
function for easily opening a ESRI shapefile by simply specifying its filename
}
\examples{
# build a spatialpoints data frame
pts = cbind(1:5, 1:5)
dimnames(pts)[[1]] = letters[1:5]
df = data.frame(a = 1:5)
row.names(df) = letters[5:1]
points <- SpatialPointsDataFrame(pts, df, match.ID = TRUE) # don't warn

Save it to a shape file
out_file  = tempfile(pattern = "test", tmpdir = tempdir(), fileext = ".shp")
mysp_object = points
writeshape(mysp_object, out_file, overw = TRUE)
}

