% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastzonal.R
\name{fastzonal}
\alias{fastzonal}
\title{fastzonal}
\usage{
fastzonal(in_rts, sp_object, start_date = NULL, end_date = NULL,
  start_band = NULL, end_band = NULL, id_field = NULL, FUN = "mean",
  out_format = "xts", small = TRUE, small_method = "centroids",
  na.rm = TRUE, verbose = FALSE)
}
\arguments{
\item{end_band}{}
}
\description{
fastzonal
}
\examples{

}

