% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates_conversion.R
\name{lb_datetodoy}
\alias{lb_datetodoy}
\title{Title lb_datetodoy}
\usage{
lb_datetodoy(dates = dates)
}
\arguments{
\item{dates}{dates to be converted. class = date, posIx, etc....}
}
\description{
Title lb_datetodoy
}

