% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_datesconversion.R
\name{lb_doytodate}
\alias{lb_doytodate}
\title{Title lb_doytodate}
\usage{
lb_doytodate(doys = doys, year = NULL)
}
\arguments{
\item{doys}{to be converted}

\item{year}{year to be used as basis (careful with LEAP years !!!)}
}
\description{
Title lb_doytodate
}
\examples{
a
}

