% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyconversion.R
\name{lb_fuzzyconversion}
\alias{lb_fuzzyconversion}
\title{lb_fuzzyconversion}
\usage{
lb_fuzzyconversion(in_raster, thresholds, ...)
}
\arguments{
\item{in_raster}{input raster file. Multiband is allowed}

\item{thresholds}{thresholds to be used for fuzzyfication. Input raster is converted to a 0:1 fuzzy score
according to the specified thresholds,}

\item{...}{any other parameter passable to "calibrate"}
}
\value{
raster object in 0:1 range obtained applying a fuzzification on the input values according to specified thresholds.
See help for "calibrate" function in QCA library for details.
}
\description{
lb_fuzzyconversion
}

