% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_theme_bw.R
\name{lb_theme_bw}
\alias{lb_theme_bw}
\title{lb_theme_bw}
\usage{
lb_theme_bw(x_ang = NULL, y_ang = NULL, xsize_t = NULL, ysize_t = NULL,
  xsize = NULL, ysize = NULL, tsize = NULL)
}
\arguments{
\item{x_ang}{angle for x labels}

\item{y_ang}{angle for y labels}

\item{xsize_t}{size for x title}

\item{ysize_t}{size for y title}

\item{xsize}{size for x labels}

\item{ysize}{size for y labels}

\item{tsize}{size for title}
}
\description{
lb_theme_bw
}

